backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.9.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
apscheduler==3.10.4
pytz==2024.2
httpx==0.27.2
python-multipart==0.0.12
pyzk==0.9.0
